Retry read for a short period of time
Even though the create and update operations are retried, it seems like
it can happen that this `DescribeInstances` fails with
`InvalidInstanceID.NotFound`. The behaviour is slightly different when
this happens. The VM is created and terraform will succeed, yet the ID
is cleared and would not be returned in outputs.

```
13:06:38 aws_instance.test[0]: Creation complete after 23s
```

vs.

```
13:06:45 aws_instance.test[1]: Creation complete after 30s (ID: i-07cae8c777c189f80)
```

Retrying this might help the API hit another server that has this data.